More work on db related stuff.
